Improve target management
Support target linking for all ways of using nifti targets (namespaced
with "NIFTI::" or not, prefixed or not, nifti1 or nifti2 file format).

More support is added for builds driving this one that use targets
namespaced with GIFTI and are not superbuilds (using FetchContent
instead of ExternalProject functionality). A namespaced alias for each
library is added during the build using the `add_gifti_library`
function. This functin, in combination with the global property
"installed_targets", is used to once again support the functionality of
the GIFTI_INSTALL_NO_LIBRARIES variable used by the ITK build.

Reverts to using GIFTICLIB_NAME for giftiio for ITK build compatibility.

Headers are set as target properties: PUBLIC_HEADER is used for
installed headers or PRIVATE_HEADER is used for headers not installed.
Support for GIFTI_INSTALL_NO_DEVELOPMENT is still supported, albeit
without if-code-blocks.

GIFTI_BUILD_APPLICATIONS prevents compilation of binary executables,
though this makes most tests fail and so most like
GIFTI_NO_INSTALL_APPLICATIONS is a better setting to use.