Add default run path for installation
Based in finding the relative path to the installation directory for
libraries this should just work. A test is added for this behavior.

One caveat might be that nifti libraries were found installed at build
time but are not installed into the same library directory. This
situation can be remedied by configuring cmake with
`-DCMAKE_INSTALL_RPATH_USE_LINK_PATH=ON`.

This has not been tested on Windows