Tidy up dockerfile
Install the nifti dependency manually so that ubuntu bionic's
default cmake version can be used. This can be changed to an
apt install at a later date.

Install cmake 3.11 to test the fetch functionality of the cmake
build (fetching the nifti dependency and test data). Add an example
of such usage.

Tidy indentation.